This is my Controller
Add GetMessageController with a GET endpoint

Introduce a new API controller `GetMessageController` in the `WebApplication1.Controllers` namespace. This controller is decorated with `[Route("api/[controller]")]` and `[ApiController]` attributes. It includes an action method `GetMessage` that handles HTTP GET requests to `api/GetMessage` and returns a "Hello World!" message as an `Ok` response.